requirements.txt
pandas==2.3.2
numpy==2.3.3
scikit-learn==1.8.0
scikit-survival==0.27.0
matplotlib==3.10.6
statsmodels==0.14.5
causalml==0.16.0
xgboost==3.2.0
pyyaml==6.0.2
tabulate==0.9.0



